Fix seedOnly logic

+++ src/core/structure-database.ts
@@ -10,7 +10,7 @@ const logger = getLogger(__filename);
 export async function structureDatabase(client: MongoClient, databaseName: string, databaseSpec: DatabaseSpec): Promise<Db> {
   try {
     const db = client.db(databaseSpec.alias || databaseName);
-    if (databaseSpec.dropFirst) {
+    if (databaseSpec.dropFirst && !databaseSpec.seedOnly) {
       logger.info('Dropping Database: %s', databaseSpec.alias || databaseName);
       logger.cli('--- Dropping Database: %s', databaseName + (databaseSpec.alias ? ` as ${databaseSpec.alias}` : ''));
       await db.dropDatabase();
